requirements.txt
torch
pytorch-lightning
pytorch-metric-learning
tensorboard
polars
numpy
jupyter
ipywidgets
toml
python-box